fix audit issues CVB-07 : Potential frontrunning attack add time lock contract MVB-02 | Incorrect input for function MVB-03 | Incorrect Error Message CON-06 : Should not collect negative fees CVB-05 : OLD VAULT ROUTER IS NOT REMOVED FROM ROLE_CONTROLLER WHEN SETTING NEW VAULT ROUTER GVB-01 : Useable net size calculation without 'sLong_ check MAK-01 : Input isOpen is not set for decreasePostionO MVR-04: Netbe abile to getCcolateralRangeQ， DBB-02: Inconsistent order guery condion OMB-01 : Charge iexecFees only when the order has sutficient collateral PAM-O1: Check collateralDelta input berore updating.positions CON-03: Missing Zero Acdress Validation MVB-02: Incorrect input for validsizen function MVE-03:Incorrect Error Message OMB-03: Colateral to be refunded shoud subtract PSB-02: Lack of input validation on account' for function remove- PSM-02: 'transToUser' should be zero if the collateral is insufficient to cover fees PSM-07 : Lack of sanity check for 'liqState PSM-08 : Collateral should be decreased to zero if collateral is insufficient to cover fees VRU-02: Comment about 'sell.' parameter amount does not match actual code PSM-09: Close the entire position yields a PartialLiquidation' reason CON-01: VARIABLES THAT COULD BE DECLARED AS IMMUTABLE PSM-03 QUERY THE PRICE TWICE WITHIN A FUNCTION RDB-01: Unused 'ReentrancyGuard' solved RDB-02: Unnecessary usage of 'SafeMath solved

+++ contracts/market/MarketLib.sol
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity ^0.8.17;
-
 import {MarketConfigStruct} from "./MarketConfigStruct.sol";
 import {IFeeRouter} from "../fee/interfaces/IFeeRouter.sol";
 import {IVaultRouter} from "../vault/interfaces/IVaultRouter.sol";
@@ -8,12 +7,20 @@ import {IOrderBook} from "../order/interface/IOrderBook.sol";
 import {IFeeRouter} from "../fee/interfaces/IFeeRouter.sol";
 import {IPositionBook} from "../position/interfaces/IPositionBook.sol";
 import {Order} from "../order/OrderStruct.sol";
-import {MarketPositionCallBackIntl, MarketOrderCallBackIntl} from "./interfaces/IMarketCallBackIntl.sol";
+import {MarketPositionCallBackIntl, MarketOrderCallBackIntl, MarketCallBackIntl} from "./interfaces/IMarketCallBackIntl.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "../utils/TransferHelper.sol";
 import "./MarketDataTypes.sol";
 
 library MarketLib {
+    /**
+     * @dev Withdraws fees from the specified collateral address.
+     * @param collAddr The address of the collateral token.
+     * @param _account The address of the account to receive the fees.
+     * @param fee The amount of fees to be withdrawn.
+     * @param collateralTokenDigits The number of decimal places for the collateral token.
+     * @param fr The address of the fee router.
+     */
     function feeWithdraw(
         address collAddr,
         address _account,
@@ -21,7 +28,7 @@ library MarketLib {
         uint8 collateralTokenDigits,
         address fr
     ) internal {
-        require(_account != address(0), "invalid user account");
+        require(_account != address(0), "feeWithdraw:!userAccount");
         if (fee < 0) {
             IFeeRouter(fr).withdraw(
                 collAddr,
@@ -34,6 +41,13 @@ library MarketLib {
         }
     }
 
+    /**
+     * @dev Withdraws profit and loss (PnL) from the vault.
+     * @param _account The address of the account to receive the PnL.
+     * @param pnl The amount of profit and loss to be withdrawn.
+     * @param collateralTokenDigits The number of decimal places for the collateral token.
+     * @param vr The address of the vault router.
+     */
     function vaultWithdraw(
         address /* collAddr */,
         address _account,
@@ -41,7 +55,7 @@ library MarketLib {
         uint8 collateralTokenDigits,
         address vr
     ) internal {
-        require(_account != address(0));
+        require(_account != address(0), "vaultWithdraw:!userAccount");
         if (pnl > 0) {
             IVaultRouter(vr).transferFromVault(
                 _account,
@@ -53,6 +67,14 @@ library MarketLib {
         }
     }
 
+    /**
+     * @dev Calculates the delta collateral for decreasing a position.
+     * @param isKeepLev Boolean flag indicating whether to keep leverage.
+     * @param size Current size of the position.
+     * @param dSize Delta size of the position.
+     * @param collateral Current collateral amount.
+     * @return deltaCollateral The calculated delta collateral.
+     */
     function getDecreaseDeltaCollateral(
         bool isKeepLev,
         uint256 size,
@@ -66,6 +88,13 @@ library MarketLib {
         }
     }
 
+    /**
+     * @dev Executes the necessary actions after updating a position.
+     * @param _item The update position event data.
+     * @param plugins The array of plugin addresses.
+     * @param erc20Token The address of the ERC20 token.
+     * @param market The address of the market.
+     */
     function afterUpdatePosition(
         MarketPositionCallBackIntl.UpdatePositionEvent memory _item,
         uint256 /* gasLimit */,
@@ -75,14 +104,24 @@ library MarketLib {
     ) internal {
         uint256 balanceBefore = IERC20(erc20Token).balanceOf(market);
         for (uint256 i = 0; i < plugins.length; i++) {
-            MarketPositionCallBackIntl(plugins[i]).updatePositionCallback(
-                _item
-            );
+            if (MarketCallBackIntl(plugins[i]).getHooksCalls().updatePosition) {
+                try
+                    MarketPositionCallBackIntl(plugins[i])
+                        .updatePositionCallback(_item)
+                {} catch {}
+            }
         }
         uint256 balanceAfter = IERC20(erc20Token).balanceOf(market);
         require(balanceAfter == balanceBefore, "ERC20 token balance changed");
     }
 
+    /**
+     * @dev Executes the necessary actions after updating an order.
+     * @param _item The update order inputs data.
+     * @param plugins The array of plugin addresses.
+     * @param collateralToken The address of the collateral token.
+     * @param market The address of the market.
+     */
     function afterUpdateOrder(
         MarketDataTypes.UpdateOrderInputs memory _item,
         uint256 /* gasLimit */,
@@ -92,12 +131,25 @@ library MarketLib {
     ) internal {
         uint256 balanceBefore = IERC20(collateralToken).balanceOf(market);
         for (uint256 i = 0; i < plugins.length; i++) {
-            MarketOrderCallBackIntl(plugins[i]).updateOrderCallback(_item);
+            if (MarketCallBackIntl(plugins[i]).getHooksCalls().updateOrder) {
+                try
+                    MarketOrderCallBackIntl(plugins[i]).updateOrderCallback(
+                        _item
+                    )
+                {} catch {}
+            }
         }
         uint256 balanceAfter = IERC20(collateralToken).balanceOf(market);
         require(balanceAfter == balanceBefore, "ERC20 token balance changed");
     }
 
+    /**
+     * @dev Executes the necessary actions after deleting an order.
+     * @param e The delete order event data.
+     * @param plugins The array of plugin addresses.
+     * @param erc20Token The address of the ERC20 token.
+     * @param market The address of the market.
+     */
     function afterDeleteOrder(
         MarketOrderCallBackIntl.DeleteOrderEvent memory e,
         uint256 /* gasLimit */,
@@ -107,12 +159,21 @@ library MarketLib {
     ) internal {
         uint256 balanceBefore = IERC20(erc20Token).balanceOf(market);
         for (uint256 i = 0; i < plugins.length; i++) {
-            MarketOrderCallBackIntl(plugins[i]).deleteOrderCallback(e);
+            if (MarketCallBackIntl(plugins[i]).getHooksCalls().deleteOrder) {
+                try
+                    MarketOrderCallBackIntl(plugins[i]).deleteOrderCallback(e)
+                {} catch {}
+            }
         }
         uint256 balanceAfter = IERC20(erc20Token).balanceOf(market);
         require(balanceAfter == balanceBefore, "ERC20 token balance changed");
     }
 
+    /**
+     * @dev Updates the cumulative funding rate for the market.
+     * @param positionBook The address of the position book.
+     * @param feeRouter The address of the fee router.
+     */
     function _updateCumulativeFundingRate(
         IPositionBook positionBook,
         IFeeRouter feeRouter
